requirements.txt
python-telegram-bot[webhooks]==20.7
asyncpg
redis
python-dotenv
